Updating local copy to current build

+++ code/modules/1713/marketplace.dm
@@ -90,7 +90,7 @@
 	else if (choice == "Change Orders")
 		var/list/currlist = list()
 		for (var/i = 1, i <= map.globalmarketplace.len, i++)
-			if (map.globalmarketplace[i][1] == user)
+			if (map.globalmarketplace[i][1] == user && map.globalmarketplace[i][5])
 				currlist += list(list(i,"[map.globalmarketplace[i][3]] of [map.globalmarketplace[i][2]], for [map.globalmarketplace[i][4]*10] silver."))
 		if (isemptylist(currlist))
 			user << "You have no orders on the market!"
@@ -158,20 +158,64 @@
 	not_movable = FALSE
 
 /obj/structure/stockmarket/attack_hand(var/mob/living/carbon/human/user as mob)
-	var/choice1 = WWinput(user, "What do you want to do?", "Stock Market", "Check Companies", list("Check Companies","Buy Stock","Sell Stock","Manage Company","Cancel"))
+	var/list/optlist = list("Check Companies","Buy Stock","Sell Stock","Manage Company")
+	if (user.leader && user.civilization != "none")
+		optlist += "Faction Treasury"
+	optlist += "Cancel"
+	var/choice1 = WWinput(user, "What do you want to do?", "Stock Market", "Check Companies", optlist)
 	if (choice1 == "Cancel")
 		return
 	else if (choice1 == "Check Companies")
 		var/list/tmplistc = sortTim(map.custom_company_value, /proc/cmp_numeric_dsc,TRUE)
 		var/body = "<html><head><title>Stock Market Companies</title></head><b>STOCK MARKET</b><br><br>"
 		for (var/relf in map.custom_company_nr)
-			body += "<b>[relf]</b>: [tmplistc[relf]*10] silver coins</br>"
+			var/vm_owned = 0
+			for(var/obj/structure/vending/sales/S in vending_machine_list)
+				if (S.owner == relf)
+					vm_owned++
+			body += "<b>[relf]</b>: [tmplistc[relf]*10] silver coins. [vm_owned] vending points.</br>"
 		body += {"<br>
 			</body></html>
 		"}
 
 		usr << browse(body,"window=artillery_window;border=1;can_close=1;can_resize=1;can_minimize=0;titlebar=1;size=250x450")
-
+	else if (choice1 == "Faction Treasury")
+		var/choicefac = WWinput(user, "What do you want to do?", "Faction Treasury", "Cancel", list("Set Taxes", "Withdraw Taxes","Cancel"))
+		if (choicefac == "Cancel")
+			return
+		else if (choicefac == "Set Taxes")
+			if (user.leader)
+				var/choiceinput1 = input(user, "What do you want the Sales Tax to be? (0-30%)","Sales Tax",map.custom_civs[user.civilization][9]) as num
+				if (choiceinput1 < 0)
+					choiceinput1 = 0
+				if (choiceinput1 > 30)
+					choiceinput1 = 10
+				var/choiceinput2 = input(user, "What do you want the Business Tax to be? (0-30%)","Business Tax",map.custom_civs[user.civilization][10]) as num
+				if (choiceinput2 < 0)
+					choiceinput2 = 0
+				if (choiceinput2 > 30)
+					choiceinput2 = 10
+				map.custom_civs[user.civilization][9] = choiceinput1
+				map.custom_civs[user.civilization][10] = choiceinput2
+				user << "<b>Sales Tax</b> set to [choiceinput1]%. <b>Business Tax</b> set to [choiceinput2]%."
+				return
+			else
+				user << "<span class='warning'>You do not have the permissions to do that!</span>"
+				return
+		else if (choicefac == "Withdraw Taxes")
+			if (map.custom_civs[user.civilization][4] == user)
+				if (map.custom_civs[user.civilization][5]>0)
+					var/obj/item/stack/money/goldcoin/GC = new/obj/item/stack/money/goldcoin(loc)
+					GC.amount = map.custom_civs[user.civilization][5]*2.5
+					map.custom_civs[user.civilization][5]=0
+					user << "You withdraw [GC.amount] gold coins in faction funds."
+					return
+				else
+					user << "<span class='notice'>There is no money to withdraw.</span>"
+					return
+			else
+				user << "<span class='warning'>You do not have the permissions to do that!</span>"
+				return
 	else if (choice1 == "Manage Company")
 		var/choice4 = WWinput(user, "What do you want to do?", "Stock Market", "View Members", list("View Members", "Distribute Profits", "Withdraw Profits","Cancel"))
 		if (choice4 == "Cancel")
@@ -189,8 +233,8 @@
 			return
 		if (choice4 == "View Members")
 			var/body = "<html><head><title>[custom_company]</title></head><b>STOCK MARKET</b><br><br>"
-				for(var/list/i in map.custom_company[custom_company])
-					body += "<b>[i[1]]</b> owns [i[2]]%.</br>"
+			for(var/list/i in map.custom_company[custom_company])
+				body += "<b>[i[1]]</b> owns [i[2]]%.</br>"
 			body += {"<br>
 				</body></html>
 			"}
@@ -201,7 +245,7 @@
 				var/tprof = map.custom_company_value[custom_company]
 				map.custom_company_value[custom_company] = 0
 				for (var/i in map.custom_company[custom_company])
-					i[3]+=i[2]*tprof
+					i[3]+=(i[2]/100)*tprof
 				user << "<span class='notice'>You distribute the profits of [custom_company].</span>"
 				return
 			else
@@ -210,10 +254,16 @@
 		else if (choice4 == "Withdraw Profits")
 			for (var/j in map.custom_company[custom_company])
 				if (j[1]==user && j[3]>0)
+					var/businesstax = 0
+					var/price_without_tax = j[3]
+					if (user.civilization != "none" && map.custom_civs[user.civilization])
+						businesstax = (map.custom_civs[user.civilization][10]/100)*j[3]
+						price_without_tax = j[3]-businesstax
+						map.custom_civs[user.civilization][5] += businesstax
 					var/obj/item/stack/money/goldcoin/GC = new/obj/item/stack/money/goldcoin(loc)
-					GC.amount = j[3]/0.4
+					GC.amount = price_without_tax*2.5
 					j[3] = 0
-					user << "<span class='notice'>You withdraw some profits.</span>"
+					user << "<span class='notice'>You withdraw [GC.amount*4] silver coins in profit, paying [businesstax*10] silver coins ([map.custom_civs[user.civilization][10]]%) in Business Tax to your faction.</span>"
 		return
 
 	else if (choice1 == "Buy Stock")
@@ -222,8 +272,14 @@
 			return
 		else
 			var/list/tempbuylist = list("Cancel")
+			var/found = FALSE
 			for(var/list/i in map.sales_registry)
-				tempbuylist += "[i[2]]% of [i[1]], at [i[3]*10] sc"
+				if (i[5])
+					tempbuylist += "[i[2]]% of [i[1]], at [i[3]*10] sc"
+					found = TRUE
+			if (!found)
+				user << "<span class='notice'>There are no stocks for sale!</span>"
+				return
 			var/choice3 = WWinput(user, "Which stock do you want to buy?", "Stock Market", "Cancel", tempbuylist)
 			if (choice3 == "Cancel")
 				return
@@ -238,24 +294,36 @@
 				if (istype(user.get_inactive_hand(), /obj/item/stack/money))
 					var/obj/item/stack/money/M = user.get_inactive_hand()
 					for(var/list/L in map.sales_registry)
-						if (L[1] == ord && text2num(L[2]) == ord_perc && text2num(L[3]) == ord_price)
-							if (L[4])
-								if (M.amount*M.value >= ord_price)
-									var/tma = M.amount*M.value
-									var/tmb = ord_price
-									var/tmc = (tma - tmb)/0.4
-									qdel(M)
-									var/obj/item/stack/money/goldcoin/GC = new/obj/item/stack/money/goldcoin(loc)
-									GC.amount = tmc
+						if (L[5] && L[1] == ord && text2num(L[2]) == ord_perc && text2num(L[3]) == ord_price)
+							if (M.amount*M.value >= ord_price)
+								var/tma = M.amount*M.value
+								var/tmb = ord_price
+								var/tmc = (tma - tmb)/0.4
+								qdel(M)
+								var/obj/item/stack/money/goldcoin/GC = new/obj/item/stack/money/goldcoin(loc)
+								GC.amount = tmc
+								if (GC.amount <= 0)
+									qdel(GC)
+								if (ishuman(L[4]))
 									var/mob/living/carbon/human/SELLER = L[4]
 									SELLER.transfer_stock_proc(ord,ord_perc,user)
+									L[5] = 0
 									map.sales_registry -= L
 									for(var/list/LL in map.sales_registry)
 										if (LL[1] == ord && LL[4]==SELLER)
+											LL[5] = 0
 											map.sales_registry -= LL
 								else
-									user << "<span class='notice'>You do not have enough money. You need [map.sales_registry[ord][3]*10] sc and you only have [M.amount*M.value*10] sc.</span>"
-									return
+									transfer_stock_nomob(ord,ord_perc,user)
+									L[5] = 0
+									map.sales_registry -= L
+									for(var/list/LL in map.sales_registry)
+										if (LL[1]==ord && LL[2]==ord_perc && LL[3]==ord_price && LL[4]==null)
+											LL[5] = 0
+											map.sales_registry -= LL
+							else
+								user << "<span class='notice'>You do not have enough money. You need [map.sales_registry[ord][3]*10] sc and you only have [M.amount*M.value*10] sc.</span>"
+								return
 				else
 					user << "<span class='notice'>You need to have money in your hands in order to buy stocks!</span>"
 					return
@@ -284,6 +352,6 @@
 					if (saleprice <=0)
 						return
 					else
-						map.sales_registry += list(list(choice2,compchoice_amt,saleprice/10,user))
+						map.sales_registry += list(list(choice2,compchoice_amt,saleprice/10,user,1))
 						user << "<span class='notice'>You sucessfully put up [compchoice_amt]% of [choice2] at [saleprice].</span>"
 						return
